split plugin into multiple files. update test_plugin to test valid topic names using plugin.publish rather than internal func.
