+++ waggle/plugin/__init__.py
@@ -1,0 +1,4 @@
+from .config import PluginConfig
+from .plugin import Plugin
+from .uploader import Uploader
+from .time import get_timestamp
